Sentiment Analysis
I and my team carried out sentiment analysis on tweets about different members of the Big Brother Naija 2020: Lockdown to predict who will win and who will be evicted that week.

Our rationale is that the winner and losers are based on voters of which a majority is on Twitter. We used this to know the mind of people on the program using #BBNAIJA2020.